The following changes have been made: - the main "objects" (byte, internal string, external string, word and CLI item) have been renamed to include the prefix _t. This is a common practice in the context of other C projects so, I've just introduced it here, too; - the main thread has been slightly overhauled in order to include a clean-up section that deals with heap memory release (not really needed right now but, depending on what I decide to do, it may come in handy); - several small changes to the code structure; - addition of a documentation related to the software architecture. I've made the picture with draw.io and added the raw file so that others may change it; - addition of several CLI options related to the TLS handshake (I've yet to be implement it for port 8729).

+++ include/utilities/generic.h
@@ -1,13 +1,13 @@
 /**************************************/
 /* Author: Pavlo Nykolyn              */
-/* Last modification date: 26-06-2025 */
+/* Last modification date: 16-07-2025 */
 /**************************************/
 
 #ifndef GENERIC_H
 #define GENERIC_H
 
-// idempotence checks
 #define INV_PNT  ((void*) 0)
-#define IDEM_INT(obj) ((obj) & ~0) // any precision will suffice
+// idempotence (useful for inhibiting const-related compilation system warnings)
+#define IDEM_INT(obj) (((long) (obj)) & ~0)
 
 #endif // GENERIC_H
